Create CoffeeBag
This program accepts two different units of weight used by the user and if necessary converts it to pounds. Then it asks for the weight of a bag of coffee and the number of bags sold to find the total price. At the end, it displays the total price and information used to find this price